+++ native/packages/react-native-bpk-component-button/stories.js
@@ -78,6 +78,15 @@ const generateButtonStoryForType = (type: string) => {
           <BpkButton
             large
             type={type}
+            title="With icon"
+            icon={getIconType(type)}
+            iconAlignment="leading"
+            onPress={action(`${type} button with icon clicked`)}
+            style={styles.buttonStyles}
+          />
+          <BpkButton
+            large
+            type={type}
             title="Icon only"
             icon={getIconType(type)}
             iconOnly
@@ -109,6 +118,14 @@ const generateButtonStoryForType = (type: string) => {
           type={type}
           title="With icon"
           icon={getIconType(type)}
+          onPress={action(`${type} button with icon clicked`)}
+          style={styles.buttonStyles}
+        />
+        <BpkButton
+          type={type}
+          title="With icon"
+          icon={getIconType(type)}
+          iconAlignment="leading"
           onPress={action(`${type} button with icon clicked`)}
           style={styles.buttonStyles}
         />